requirements.txt
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.36.0
accelerate>=0.25.0
peft>=0.7.0
bitsandbytes>=0.41.0
Pillow>=10.0.0
pyyaml>=6.0
omegaconf>=2.3.0
datasets>=2.14.0
rich>=13.0.0
tqdm>=4.66.0
scikit-learn>=1.3.0
numpy>=1.24.0
pandas>=2.0.0
